requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
async-generator==1.10
attrs==23.1.0
autocomplete-light==1.1.5
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudinary==1.32.0
coverage==7.2.5
cryptography==40.0.2
cssselect2==0.7.0
dj-database-url==1.3.0
Django==4.2.1
django-cloudinary-storage==0.3.0
django-crispy-forms==2.0
django-filter==23.2
django-postgresql==0.0.3
exceptiongroup==1.1.1
fontawesomefree==6.4.0
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
idna==3.4
iniconfig==2.0.0
lxml==4.9.2
oscrypto==1.3.0
outcome==1.2.0
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
psycopg2==2.9.6
pycparser==2.21
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
pypdf==3.8.1
pypng==0.20220715.0
PySocks==1.7.1
pytest==7.3.1
pytest-cov==4.0.0
pytest-splinter==3.3.2
python-bidi==0.4.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.30.0
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
splinter==0.19.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uritools==4.0.1
urllib3==1.26.15
webencodings==0.5.1
whitenoise==6.4.0
wsproto==1.2.0
xhtml2pdf==0.2.11
